server/src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
package ru.practicum.shareit.user;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.practicum.shareit.exception.DuplicatedDataException;
import ru.practicum.shareit.exception.NotFoundException;
import ru.practicum.shareit.user.dao.UserRepository;
import ru.practicum.shareit.user.dto.UserDto;
import ru.practicum.shareit.user.model.User;

import java.util.Collection;

@RequiredArgsConstructor
@Service
public class UserServiceImpl implements UserService {
    private final UserRepository userRepository;

    @Override
    public Collection<UserDto> findAll() {
        return UserMapper.toUserDtoCollection(userRepository.findAll());
    }

    @Override
    public UserDto create(UserDto userDto) {
        if (userRepository.existsByEmail(userDto.getEmail())) {
            throw new DuplicatedDataException("User with e-mail = " + userDto.getEmail() + "exist!");
        }
        return UserMapper.toUserDto(userRepository.save(UserMapper.toUser(userDto)));
    }

    @Override
    public UserDto update(Long userId, UserDto userDto) {
        User user = userRepository.findById(userId).orElseThrow(() -> {
            throw new NotFoundException("User (id = " + userId + ") not found!");
        });
        if (userDto.getEmail() != null
                && !userDto.getEmail().isBlank()) {
            if (userRepository.existsByEmail(userDto.getEmail())) {
                throw new DuplicatedDataException("User with e-mail = " + userDto.getEmail() + "exist!");
            }
            user.setEmail(userDto.getEmail());
        }
        if (userDto.getName() != null
                && !userDto.getName().isBlank()) {
            user.setName(userDto.getName());
        }
        return UserMapper.toUserDto(userRepository.save(user));
    }

    @Override
    public UserDto getUserDtoById(Long userId) {
        return UserMapper.toUserDto(userRepository.findById(userId).orElseThrow(() -> {
            throw new NotFoundException("User id = " + userId + " not found!");
        }));
    }

    @Override
    public void delete(Long userId) {
        userRepository.deleteById(userId);
    }
}
